changed method loc to helper_function.py
