+++ server/app/main.py
@@ -27,7 +27,7 @@ from server.app.query_execution_lineage_d3force import query_execution_lineage_d
 from server.app.query_execution_lineage_d3tree import query_execution_lineage_d3tree
 from server.app.query_artifact_lineage_d3tree import query_artifact_lineage_d3tree
 from server.app.query_visualization_artifact_execution import query_visualization_artifact_execution
-from cmflib.utils.execution_utils import create_unique_executions
+from cmflib.utils.helper_functions import create_unique_executions
 from cmflib.cmf_exception_handling import MlmdNotFoundOnServer
 from pathlib import Path
 import os
